lib: reorganise files
