+++ lib/ascii_charts/version.rb
@@ -1,0 +1,3 @@
+module AsciiCharts
+  VERSION = '0.9.2'
+end
